+++ source/patterns/@aws-solutions-constructs/aws-cloudfront-s3/test/test.cloudfront-s3.test.ts
@@ -141,7 +141,7 @@ test('test cloudfront with custom domain names', () => {
         Bucket: {
           "Fn::GetAtt": [
             "testcloudfronts3CloudfrontLoggingBucket985C0FE8",
-            "DomainName"
+            "RegionalDomainName"
           ]
         }
       },