+++ source/patterns/@aws-solutions-constructs/core/index.ts
@@ -43,3 +43,7 @@ export * from './lib/elasticsearch-helper';
 export * from './lib/step-function-defaults';
 export * from './lib/step-function-helper';
 export * from './lib/dynamodb-table-helper';
+export * from './lib/vpc-defaults';
+export * from './lib/vpc-helper';
+export * from './lib/sagemaker-defaults';
+export * from './lib/sagemaker-helper';