+++ source/patterns/@aws-solutions-constructs/core/test/cloudfront-distribution-s3-helper.test.ts
@@ -162,7 +162,7 @@ test('test cloudfront with no security headers ', () => {
         Bucket: {
           "Fn::GetAtt": [
             "CloudfrontLoggingBucket3C3EFAA7",
-            "DomainName"
+            "RegionalDomainName"
           ]
         }
       },
@@ -233,7 +233,7 @@ test('test cloudfront override cloudfront custom domain names ', () => {
         Bucket: {
           "Fn::GetAtt": [
             "CloudfrontLoggingBucket3C3EFAA7",
-            "DomainName"
+            "RegionalDomainName"
           ]
         }
       },
@@ -308,7 +308,7 @@ test('test cloudfront override cloudfront logging bucket ', () => {
         Bucket: {
           "Fn::GetAtt": [
             "loggingbucket6D73BD53",
-            "DomainName"
+            "RegionalDomainName"
           ]
         }
       },
@@ -392,7 +392,7 @@ test('test cloudfront override properties', () => {
           Bucket: {
             "Fn::GetAtt": [
               "CloudfrontLoggingBucket3C3EFAA7",
-              "DomainName"
+              "RegionalDomainName"
             ]
           }
         },