+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.70.0] - 2020-10-26
+
+### Added
+- aws-events-rule-kinesisstreams pattern added ([#59](https://github.com/awslabs/aws-solutions-constructs/issues/59))
+- aws-events-rule-kinesisfirehose-s3 pattern added ([#72](https://github.com/awslabs/aws-solutions-constructs/issues/72))
+- aws-lambda-sagemaker pattern added ([#23](https://github.com/awslabs/aws-solutions-constructs/issues/23))
+
+### Changed
+- Upgraded all patterns to CDK v1.70.0
+
 ## [1.69.0] - 2020-10-21
 
 ### Changed